Move documents.py and session_view.py up one level
